global effects provider 5/5: extract show/hideEffect part of provider to specialized class
