+++ src/effects/effects_base/internal/effectexecutionscenario.cpp
@@ -282,7 +282,7 @@ muse::Ret EffectExecutionScenario::doPerformEffect(au3::Au3Project& project, con
         if (effect->IsInteractive() && (flags& EffectManager::kConfigured) == 0) {
             const auto access = std::make_shared<SimpleEffectSettingsAccess>(*settings);
             EffectInstanceId instanceId = effectInstancesRegister()->regInstance(effectId, pInstanceEx, access);
-            muse::Ret ret = effectsProvider()->showEffect(effectId, instanceId);
+            muse::Ret ret = effectViewController()->showEffect(effectId, instanceId);
             effectInstancesRegister()->requestUpdateSettings(instanceId);
             effectInstancesRegister()->unregInstance(instanceId);
             if (ret) {
