+++ src/effects/effects_base/internal/effectsprovider.cpp
@@ -133,67 +133,3 @@ Effect* EffectsProvider::effect(const EffectId& effectId) const
 
     return effect;
 }
-
-namespace {
-IEffectViewLauncherPtr getLauncher(const EffectId& effectId, const IEffectViewLaunchRegister& launchRegister)
-{
-    PluginID pluginID = effectId.toStdString();
-    const PluginDescriptor* plug = PluginManager::Get().GetPlugin(pluginID);
-    if (!plug || !PluginManager::IsPluginAvailable(*plug)) {
-        LOGE() << "plugin not available, effectId: " << effectId;
-        return {};
-    }
-
-    const auto family = au::au3::wxToStdString(plug->GetEffectFamily());
-    const auto launcher = launchRegister.launcher(family);
-    IF_ASSERT_FAILED(launcher) {
-        LOGE() << "not found launcher for family:" << family;
-        return {};
-    }
-    return launcher;
-}
-
-void callOnLauncher(const RealtimeEffectStatePtr& state, const IEffectViewLaunchRegister& launchRegister,
-                    std::function<void(const IEffectViewLauncher&, const RealtimeEffectStatePtr&)> func)
-{
-    IF_ASSERT_FAILED(state) {
-        return;
-    }
-    if (const auto launcher = getLauncher(au::au3::wxToString(state->GetID()), launchRegister)) {
-        func(*launcher, state);
-    }
-}
-}
-
-muse::Ret EffectsProvider::showEffect(const EffectId& effectId, const EffectInstanceId& instanceId)
-{
-    LOGD() << "try open effect: " << effectId << ", instanceId: " << instanceId;
-
-    if (!loadEffect(effectId)) {
-        return muse::make_ret(muse::Ret::Code::NotSupported);
-    }
-
-    const auto launcher = getLauncher(effectId, *viewLaunchRegister());
-    if (!launcher) {
-        return muse::make_ret(muse::Ret::Code::NotSupported);
-    }
-
-    Ret ret = launcher->showEffect(instanceId);
-
-    LOGD() << "open ret: " << ret.toString();
-    return ret;
-}
-
-void EffectsProvider::showEffect(const RealtimeEffectStatePtr& state) const
-{
-    callOnLauncher(state, *viewLaunchRegister(), [](const IEffectViewLauncher& launcher, const RealtimeEffectStatePtr& state) {
-        launcher.showRealtimeEffect(state);
-    });
-}
-
-void EffectsProvider::hideEffect(const RealtimeEffectStatePtr& state) const
-{
-    callOnLauncher(state, *viewLaunchRegister(), [](const IEffectViewLauncher& launcher, const RealtimeEffectStatePtr& state) {
-        launcher.hideRealtimeEffect(state);
-    });
-}
