+++ src/effects/effects_base/internal/effectviewcontroller.h
@@ -1,0 +1,32 @@
+/*
+ * Audacity: A Digital Audio Editor
+ */
+#pragma once
+
+#include "framework/global/modularity/ioc.h"
+
+#include "../ieffectviewlaunchregister.h"
+
+#include "../ieffectviewcontroller.h"
+#include "ieffectsprovider.h"
+
+class EffectBase;
+class EffectSettingsAccess;
+class TrackList;
+
+namespace au::effects {
+class EffectViewController : public IEffectViewController, public muse::Contextable
+{
+    muse::GlobalInject<IEffectsProvider> effectsProvider;
+    muse::Inject<IEffectViewLaunchRegister> viewLaunchRegister{ this };
+
+public:
+    EffectViewController(const muse::modularity::ContextPtr& ctx)
+        : muse::Contextable(ctx) {}
+
+    muse::Ret showEffect(const EffectId& effectId, const EffectInstanceId& instanceId) override;
+
+    void showEffect(const RealtimeEffectStatePtr& state) const override;
+    void hideEffect(const RealtimeEffectStatePtr& state) const override;
+};
+}
