+++ src/projectscene/view/trackspanel/realtimeeffectlistitemmodel.h
@@ -7,6 +7,7 @@
 #include "async/asyncable.h"
 #include "effects/effects_base/ieffectsprovider.h"
 #include "effects/effects_base/irealtimeeffectservice.h"
+#include "effects/effects_base/ieffectviewcontroller.h"
 
 #include <QObject>
 
@@ -19,6 +20,7 @@ class RealtimeEffectListItemModel : public QObject, public muse::Contextable, pu
 
     muse::GlobalInject<effects::IEffectsProvider> effectsProvider;
     muse::ContextInject<effects::IRealtimeEffectService> realtimeEffectService{ this };
+    muse::ContextInject<effects::IEffectViewController> effectViewController{ this };
 
 public:
     RealtimeEffectListItemModel(QObject* parent, effects::RealtimeEffectStatePtr effectState);
